refine new launch logic and add functions to firework to reset launch
